hw_2/requirements.txt
novokshanovLatexGen==0.4.0
